Aim slightly right or left of the centre of ship
This is a minor change in the autoplay algorithm to avoid aiming
directly at the centre of the ship. The problem with aiming directly
at the centre of the ship is that the boulders are released from the
centre of the ship, so aiming directly at the centre often leads to
hitting a recently released boulder instead of the ship. This can lead
to loss of valuable time especially when the ships are too close to
the player. As a result, this change ensures that pulses are aimed
slightly to the left or right of the ship to avoid hitting recently
released boulders.